refactor(highlight): use class to encapsulate global state
