+++ src/constants/highlight.ts
@@ -1,0 +1,5 @@
+export const commandIds = {
+  toggleEnabled: 'veco.highlight.toggleEnabled',
+  listAnnotations: 'veco.highlight.listAnnotations',
+  showOutputChannel: 'veco.highlight.showOutputChannel',
+} as const
